Create a file for the base classes for Formatters
